fix params id

+++ controllers/user/editProfile.js
@@ -8,6 +8,9 @@ module.exports = {
   async editProfile(req, res) {
     try {
       const {
+        params: {
+          id,
+        },
         user,
         files: {
           picture,
@@ -18,49 +21,52 @@ module.exports = {
         },
       } = req;
 
-      const date = Date.now();
-      const pictureName = `images/${date}${picture.name}`;
+      if (+id === +user.id) {
+        const date = Date.now();
+        const pictureName = `images/${date}${picture.name}`;
 
-      const payload = {};
+        const payload = {};
 
-      const userLogin = login ? login.trim() : '';
-      const userEmail = email ? email.trim() : '';
+        const userLogin = login ? login.trim() : '';
+        const userEmail = email ? email.trim() : '';
 
-      if (userLogin !== '') {
-        payload.login = userLogin;
-      }
-
-      if (userEmail !== '') {
-        payload.email = userEmail;
-      }
-
-      if (picture) {
-        try {
-          await unlink(`public/${user.picture}`);
-          await picture.mv(`public/${pictureName}`);
-          payload.picture = pictureName;
-        } catch (error) {
-          return res
-            .status(RESPONSE_STATUSES.NOT_FOUND)
-            .send(ERROR_MESSAGE.NOT_FOUND);
+        if (userLogin !== '') {
+          payload.login = userLogin;
         }
+
+        if (userEmail !== '') {
+          payload.email = userEmail;
+        }
+
+        if (picture) {
+          try {
+            await unlink(`public/${user.picture}`);
+            await picture.mv(`public/${pictureName}`);
+            payload.picture = pictureName;
+          } catch (error) {
+            return res
+              .status(RESPONSE_STATUSES.NOT_FOUND)
+              .send(ERROR_MESSAGE.NOT_FOUND);
+          }
+        }
+
+        await Users.update(
+          payload,
+          { where: { id } },
+        );
+        const editedUser = await Users.findOne(
+          {
+            attributes: ['login', 'id', 'email', 'firstName', 'lastName', 'picture'],
+            where: { id },
+          },
+        );
+        return res
+          .status(RESPONSE_STATUSES.OK)
+          .send({ user: editedUser });
       }
-
-      await Users.update(
-        payload,
-        { where: { id: user.id } },
-      );
-
-      const editedUser = await Users.findOne(
-        {
-          attributes: ['login', 'id', 'email', 'firstName', 'lastName', 'picture'],
-          where: { id: user.id },
-        },
-      );
-
       return res
-        .status(RESPONSE_STATUSES.OK)
-        .send({ user: editedUser });
+        .status(RESPONSE_STATUSES.BAD_REQUEST)
+        .send(ERROR_MESSAGE.INVALID_ID);
     } catch (error) {
       return res
         .status(RESPONSE_STATUSES.INTERNAL_SERVER_ERROR)
